Move fXstream template into its own file
